Update replaceIPaddresses
It prints the output to a file and replaces the source and destination IP addresses before writing to the file. It replaces some of the unknown IP addresses with unknown.

Next steps: some of the nfdump lines are still missing from the output file so this problem needs to be solved.