Update and rename convertPCAP to InitialDataProcessing
The file uses nfpcapd to convert pcap files into nfcapd format. Nfdump will process the flows and output a txt file. This is processed by the python script to extract the IP addresses to be used with nslookup. The nfdump file and nslookup file is read by the mergeFiles python script which sends the output to the database.